create new map model, set set lat and lng, and display a map on the show page. still some bugs that require refreshing page

+++ db/migrate/20140222185249_create_maps.rb
@@ -2,8 +2,8 @@ class CreateMaps < ActiveRecord::Migration
   def change
     create_table :maps do |t|
       t.string :address
-      t.decimal :lat
-      t.decimal :lng
+      t.float :lat
+      t.float :lng
 
       t.timestamps
     end
